feat(server): port the data layer from D1/R2/KV to Postgres and S3
The substantive half of the Docker port. The Drizzle schema moves from
SQLite to Postgres, and the three Cloudflare bindings become ordinary
services built once at startup: a Bun SQL pool, an S3 client, and a
counter table.

Schema (34 tables + 1 new):
  - sqliteTable -> pgTable throughout; a ts() column factory replaces the
    repeated timestamp options, matching the file's existing idiom.
  - Every instant is timestamptz. Drizzle maps both it and the old epoch-ms
    integers to a JS Date, so application code is untouched by the change
    while the database gets a column it can reason about.
  - measurement_log.value and medicine_log.amount are doublePrecision, NOT
    real. SQLite's REAL is an 8-byte double; Postgres' `real` is 4-byte
    single precision and silently stored a weight of 8.4 as
    8.399999618530273. Caught by the stats suite.
  - New rate_limit table replacing the KV namespace.
  - The 15 SQLite migrations are replaced by one generated Postgres
    baseline. No production data is being carried over, so a hand-written
    migration chain would be fiction; the old files remain in git history.

Transactions replace every D1 batch() workaround. D1 had no transactions,
which is why contact/calendar writes were split into a batch plus a
separate ownership check, and why invite redemption was hand-written SQL
with duplicated guards and a compensating DELETE. Each is now one
transaction, and redeem takes SELECT … FOR UPDATE — so max_uses is
genuinely enforced under concurrency rather than merely hard to exceed.

Three dialect traps the type checker could not have caught:
  - isUniqueViolation() matched the literal "UNIQUE", which is SQLite's
    wording. Postgres says "duplicate key value violates unique
    constraint", so the "baby already has an active session" path would
    have degraded to a 500. Now matches SQLSTATE 23505.
  - Admin family counts used sql<number>`COUNT(*)`; Postgres COUNT is
    bigint and the driver returns bigints as STRINGS, so the API served
    "1" where its own schema promises a number. Cast with ::int.
  - Calendar reminder due-ness was epoch-millisecond arithmetic
    (start_time - minutes * 60000). `timestamptz - integer` is not an
    operator Postgres defines; it is interval arithmetic now.

Storage deliberately accepts Blob | string and NOT a ReadableStream. Bun's
S3 client does not reject a stream — it writes the literal string
"[object ReadableStream]" and reports success. The R2 code passed
file.stream(), so accepting one here would have made silent upload
corruption both easy and invisible. Verified against MinIO directly.

The rate limiter gains a real fix rather than a translation. It read
cf-connecting-ip, which Cloudflare guarantees; behind an arbitrary ingress
that header is absent and every caller would have shared one bucket,
quietly disabling the brute-force brake. It now reads X-Forwarded-For only
when TRUSTED_PROXY_HOPS says how many proxies to trust, counting from the
right, and falls back to the peer address. The KV counter's read-then-write
race is gone too: one INSERT … ON CONFLICT … RETURNING is atomic, which
matters once replicas share a database.

Services are built ONCE (services.ts), memoized on the Env object's
identity. On Workers they had to be per-request because bindings only
existed inside the handler, so every single request constructed a Stripe
client and the entire better-auth plugin chain.

Also here: main.ts (Bun.serve, static assets, SPA fallback, the CSP that
the Cloudflare _headers file used to supply, graceful shutdown), cron.ts +
cron-cli.ts replacing the Workers scheduled() handler, /healthz + /readyz,
and robots.txt/sitemap.xml moved from build-time generation to runtime so
one image can be promoted from test to production.

Tests move from vitest + @cloudflare/vitest-pool-workers to `bun test`
against a real Postgres, with an in-memory Storage. All 187 original tests
pass, plus 13 new config tests: 200 green, and 23s instead of 81s.

Phase: Docker/Postgres port.

+++ src/server/routes/admin.ts
@@ -144,8 +144,12 @@ export const adminApp = createApp<AppEnv>()
         slug: schema.organization.slug,
         plan: schema.organization.plan,
         createdAt: schema.organization.createdAt,
-        members: sql<number>`(SELECT COUNT(*) FROM member WHERE member.organization_id = ${schema.organization.id})`,
-        babies: sql<number>`(SELECT COUNT(*) FROM baby WHERE baby.family_id = ${schema.organization.id})`,
+        // ::int is load-bearing. COUNT() is bigint in Postgres, and the
+        // driver hands bigints back as STRINGS to avoid losing precision —
+        // so without the cast these came out as "3" and the sql<number>
+        // annotation quietly became a lie the API then served.
+        members: sql<number>`(SELECT COUNT(*)::int FROM member WHERE member.organization_id = ${schema.organization.id})`,
+        babies: sql<number>`(SELECT COUNT(*)::int FROM baby WHERE baby.family_id = ${schema.organization.id})`,
         lastFeedMs: max(schema.feedLog.time),
       })
       .from(schema.organization)
